Let's fix price?

+++ src/main/java/net/danh/mmoshop/Data/Item.java
@@ -67,7 +67,7 @@ public class Item {
         int a = getPlayerAmount(p, item);
         if (a >= amount) {
             removeItems(p, item, amount);
-            ExecuteCommand(p, Command, (int) (price * amount));
+            ExecuteCommand(p, Command, (price * amount));
             sendPlayerMessage(p, Objects.requireNonNull(getLanguage().getString("SELL_ITEMS")).replaceAll("%symbol%", Matcher.quoteReplacement(symbol)).replaceAll("%item%", Objects.requireNonNull(item.getItemMeta()).getDisplayName()).replaceAll("%price%", String.valueOf(price * amount)).replaceAll("%amount%", String.format("%,d", amount)));
         } else {
             sendPlayerMessage(p, Objects.requireNonNull(getLanguage().getString("NOT_ENOUGH_ITEM")).replaceAll("%symbol%", Matcher.quoteReplacement(symbol)).replaceAll("%item%", Objects.requireNonNull(item.getItemMeta()).getDisplayName()));
@@ -90,11 +90,11 @@ public class Item {
         return null;
     }
 
-    public static Integer Cost(Player p, String placeholders) {
-        return BigDecimal.valueOf(Long.parseLong(PlaceholderAPI.setPlaceholders(p, placeholders))).intValue();
+    public static Double Cost(Player p, String placeholders) {
+        return BigDecimal.valueOf(Long.parseLong(PlaceholderAPI.setPlaceholders(p, placeholders))).doubleValue();
     }
 
-    public static void ExecuteCommand(Player p, List<String> commands, Integer cost) {
+    public static void ExecuteCommand(Player p, List<String> commands, Double cost) {
         for (String cmd : commands) {
             if (cmd.startsWith("[CMD] ")) {
                 String command = PlaceholderAPI.setPlaceholders(p, cmd.replace("[CMD] ", "").replaceAll("%cost%", String.valueOf(cost)));
@@ -114,8 +114,8 @@ public class Item {
             return;
         }
         item.setAmount(amount);
-        if (Cost(p, Placeholder) >= (int) (price * amount)) {
-            ExecuteCommand(p, commands, (int) (price * amount));
+        if (Cost(p, Placeholder) >= (price * amount)) {
+            ExecuteCommand(p, commands, (price * amount));
             p.getInventory().addItem(item);
             sendPlayerMessage(p, Objects.requireNonNull(getLanguage().getString("BUY_ITEMS")).replaceAll("%symbol%", Matcher.quoteReplacement(symbol)).replaceAll("%item%", Objects.requireNonNull(item.getItemMeta()).getDisplayName()).replaceAll("%price%", String.valueOf(price * amount)).replaceAll("%amount%", String.format("%,d", amount)));
         } else {
